pdf-extract-backend/services/img_desc_service/requirements.txt
Flask==2.2.5
Pillow==11.1.0
protobuf==6.30.0
python-dotenv==1.0.1
google-genai